Edit input handling and output formatting
Rework Duke class to delegate input parsing responsibilities entirely to
the Parser class, enhancing separation of concerns and adhering more
closely to OOP principles. This change ensures that input parsing logic
is at one place, making the code easier to understand and maintain.

Modify the run method in Duke to adjust the timing and conditions under
which ui.printLine() is invoked. This prevents the double line print
issue at the end of the program execution, leading to cleaner and more
user-friendly output.

These adjustments improve the overall design and usability, improving
the user interaction experience by providing a clearer output format.
